fix: move the NewsType import
